+++ test/libsolidity/semanticTests/array/array_copy_cleanup_uint128.sol
@@ -19,5 +19,7 @@ contract C {
         return true;
     }
 }
+// ====
+// compileViaYul: also
 // ----
 // f() -> true