+++ test/libsolidity/semanticTests/array/inline_array_return.sol
@@ -11,5 +11,7 @@ contract C {
     }
 }
 
+// ====
+// compileViaYul: also
 // ----
 // f() -> 1, 2, 3, 4, 5