+++ test/libsolidity/semanticTests/constructor/constructor_static_array_argument.sol
@@ -8,6 +8,8 @@ contract C {
     }
 }
 
+// ====
+// compileViaYul: also
 // ----
 // constructor(): 1, 2, 3, 4 ->
 // a() -> 1