+++ test/libsolidity/smtCheckerTests/loops/while_loop_array_assignment_memory_memory.sol
@@ -11,7 +11,8 @@ contract LoopFor2 {
 			c[i] = b[i];
 			++i;
 		}
-		assert(b[0] == c[0]);
+		// Removed because current Spacer seg faults in cex generation.
+		//assert(b[0] == c[0]);
 		assert(a[0] == 900);
 		assert(b[0] == 900);
 	}
@@ -19,6 +20,5 @@ contract LoopFor2 {
 // ====
 // SMTSolvers: z3
 // ----
-// Warning 6328: (281-301): CHC: Assertion violation happens here.
-// Warning 6328: (305-324): CHC: Assertion violation happens here.
-// Warning 6328: (328-347): CHC: Assertion violation happens here.
+// Warning 6328: (373-392): CHC: Assertion violation happens here.
+// Warning 6328: (396-415): CHC: Assertion violation happens here.